drop support for Python 2.6
Because we will need memoryview for efficient implementation of handling
writes (which will be in thousands and every one at least half a kiB
in size), we can't support Python 2.6.

RHEL 7, which has the interesting for us btrfs file system, has python 2.7
in standard installation, so it's not a problem there either.